+++ Concurrency.Implementation/Logging/LogFormat.cs
@@ -9,12 +9,12 @@ namespace Concurrency.Implementation.Logging
     [Serializable]
     class LogFormat<TState>
     {
-        public long sequenceNumber;
+        public int sequenceNumber;
         public LogType logRecordType;
         public Guid coordinatorKey;
-        public long txn_id;
+        public int txn_id;
         public TState state;
-        public LogFormat(long sequenceNumber, LogType logRecordType, Guid coordinatorKey, long txn_id, TState state)
+        public LogFormat(int sequenceNumber, LogType logRecordType, Guid coordinatorKey, int txn_id, TState state)
         {
             this.sequenceNumber = sequenceNumber;
             this.logRecordType = logRecordType;
@@ -23,7 +23,7 @@ namespace Concurrency.Implementation.Logging
             this.state = state;
         }
 
-        public LogFormat(long sequenceNumber, LogType logRecordType, Guid coordinatorKey, long txn_id)
+        public LogFormat(int sequenceNumber, LogType logRecordType, Guid coordinatorKey, int txn_id)
         {
             this.sequenceNumber = sequenceNumber;
             this.logRecordType = logRecordType;
